+++ packages/agent-tools/src/index.ts
@@ -8,6 +8,7 @@
  */
 
 export * from "./capabilities.js";
+export * from "./edit.js";
 export * from "./glob.js";
 export * from "./tool.js";
 export * from "./registry.js";